Add dry run script and refactor scaffold page functionality
