+++ src/sample.test.ts
@@ -1,0 +1,5 @@
+import { expect, test } from "vitest";
+
+test("should pass", () => {
+  expect(true).toBe(true);
+});
